scripts/requirements.txt
pyyaml>=6.0
jsonschema>=4.18.0
jinja2>=3.0
pytest>=7.0
pytest-cov>=4.0
